+++ src/app/[assembly]/[entityType]/[entityID]/[[...tab]]/_GeneTabs/_Gene/CalcNearbyCCREs.tsx
@@ -13,6 +13,8 @@ import {
   Typography,
   Tooltip,
 } from "@mui/material";
+import { formatTssOffset } from "./utils";
+
 interface CalculateNearbyCCREsPopperProps {
   open: boolean;
   anchorEl: HTMLButtonElement | null;
@@ -92,12 +94,6 @@ const CalculateNearbyCCREsPopper: React.FC<CalculateNearbyCCREsPopperProps> = ({
       </ClickAwayListener>
     </Popper>
   );
-};
-
-// Signed offset relative to the TSS: negative = upstream, positive = downstream
-export const formatTssOffset = (value: number) => {
-  if (value === 0) return "TSS";
-  return `${Math.abs(value) / 1000}kb ${value < 0 ? "upstream" : "downstream"}`;
 };
 
 const tssMarks = [
