+++ src/app/[assembly]/[entityType]/[entityID]/[[...tab]]/_GeneTabs/_cCREs/DistanceLinkedCcres.tsx
@@ -6,7 +6,8 @@ import { LinkComponent } from "common/components/LinkComponent";
 import { Table, TableColDef } from "@weng-lab/ui-components";
 import CalculateIcon from "@mui/icons-material/Calculate";
 import { useMemo, useState, useRef } from "react";
-import CalculateNearbyCCREsPopper, { formatTssOffset } from "../_Gene/CalcNearbyCCREs";
+import CalculateNearbyCCREsPopper from "../_Gene/CalcNearbyCCREs";
+import { formatTssOffset } from "../_Gene/utils";
 import { Assembly, GenomicRange } from "common/types/globalTypes";
 import { InfoOutlineRounded } from "@mui/icons-material";
 import { calcDistCcreToTSS } from "common/utils";
