+++ src/common/OpenEntitiesContext/context.ts
@@ -1,0 +1,6 @@
+"use client";
+
+import { createContext, Dispatch } from "react";
+import { OpenEntityAction, OpenEntityState } from "./types";
+
+export const OpenEntitiesContext = createContext<[OpenEntityState, Dispatch<OpenEntityAction>]>(null);
