+++ src/common/OpenEntitiesContext/index.ts
@@ -3,4 +3,5 @@ export * from "./guards";
 export * from "./reducer";
 export * from "./helpers";
 // export these individually so that "use client" directive does not get applied to this file
-export { OpenEntitiesContext, OpenEntitiesContextProvider } from "./context";
+export { OpenEntitiesContext } from "./context";
+export { OpenEntitiesContextProvider } from "./provider";
