+++ src/common/components/MainSearchAutocomplete/index.ts
@@ -1,0 +1,2 @@
+export * from "./MainSearchAutocomplete";
+export * from "./mainSearchHelpers";
